Set standard byte order and size on unpack
We should use the standard size and little-endian. Realistically, the
only systems this will be run on are x86, x86_64, or ARM which are all
modern architectures using little-endian encoding. We should also make
sure we are using standard sizes. Unsigned longs on Windows are 4 bytes
but on other sytems they are 8, so we should use the standard sizing,
not the native system sizing as this isn't portable.